requirements.txt
flask_admin
flask_apscheduler
flask_markdown
flask_security
flask_sqlalchemy
flask_wtf
flask.ext
flask==2.3.1
lxml
packaging
pymysql
pytz
pyyaml
sqlalchemy==1.4.45
watchdog
wtforms_sqlalchemy
gunicorn
python-ldap
cssutils
cryptography